requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.40.0
google-generativeai==0.7.2
Pillow==10.4.0
Werkzeug==3.0.3
